Add NPM packaging with rollup.
