+++ .eslintrc.js
@@ -1,0 +1,4 @@
+module.exports = {
+  extends: ['@foresee/eslint-config-foresee'],
+  ignorePatterns: ['dist/'],
+};
