+++ .gitignore
@@ -1,0 +1,3 @@
+dist/
+node_modules/
+.eslintcache
