+++ .prettierrc.js
@@ -1,0 +1,3 @@
+module.exports = {
+  ...require('@foresee/eslint-config-foresee/.prettierrc.js'),
+};
