+++ package.json
@@ -1,0 +1,43 @@
+{
+  "name": "backbone-oauth-1.0a",
+  "version": "0.10.0",
+  "description": "OAuth 1.0 implementation for Backbone",
+  "main": "dist/index.js",
+  "scripts": {
+    "build": "rollup -c",
+    "lint": "TIMING=1 eslint --ext .js . --cache",
+    "pre-commit": "npm run lint && npm test"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/foreseecode/backbone-oauth-1.0a.git"
+  },
+  "keywords": [
+    "backbone",
+    "oauth"
+  ],
+  "author": "ForeSee",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/foreseecode/backbone-oauth-1.0a/issues"
+  },
+  "homepage": "https://github.com/foreseecode/backbone-oauth-1.0a#readme",
+  "devDependencies": {
+    "@babel/core": "^7.10.5",
+    "@babel/preset-env": "^7.10.4",
+    "@foresee/eslint-config-foresee": "^1.3.0",
+    "@rollup/plugin-babel": "^5.1.0",
+    "@rollup/plugin-node-resolve": "^8.4.0",
+    "eslint": "^7.5.0",
+    "eslint-plugin-prettier": "^3.1.4",
+    "husky": "^4.2.5",
+    "prettier": "^2.0.5",
+    "rollup": "^2.23.0",
+    "rollup-plugin-uglify": "^6.0.4"
+  },
+  "husky": {
+    "hooks": {
+      "preCommit": "npm run pre-commit"
+    }
+  }
+}
